+++ app/src/protyle/publish/index.ts
@@ -3,8 +3,8 @@ import {lockFile} from "../../dialog/processSystem";
 import {hideMessage, showMessage} from "../../dialog/message";
 import {PrintToPDFOptions} from "electron";
 import PUBLISH_TYPE_CONSTANTS, {getApiParams} from "./util";
-import wordpress_compatible from "./wordpress/lib/wordpress-compatible";
 import metaWeblogApiClient from "./metaweblog/metaweblog-api-client";
+import wordpressApiClient from "./wordpress/wordpress-api-client";
 
 // publishMdContent
 export const publishMdContent = (id: string, type: string, meta: any, callback: Function) => {
@@ -49,7 +49,12 @@ const doPublish = (id: string, type: string, meta: any, content: any) => {
         [postidKey]: "99999",
     };
 
-    // wordpressApi_getPosts(apiParams);
+    const wordpressApi = wordpressApiClient(PUBLISH_TYPE_CONSTANTS.API_TYPE_WORDPRESS);
+    const result2 = wordpressApi.getPosts(10);
+    // @ts-ignore
+    result2.then(function (reslove: any, reject: any) {
+        console.log("wordpress getPosts=>", reslove);
+    });
 
     const metaWeblogApi = metaWeblogApiClient(PUBLISH_TYPE_CONSTANTS.API_TYPE_WORDPRESS);
     const result = metaWeblogApi.getRecentPosts(10);
@@ -66,31 +71,3 @@ const doPublish = (id: string, type: string, meta: any, content: any) => {
         // console.log("doPublish content=>", content);
     });
 };
-
-async function wordpressApi_getPosts(apiParams: any) {
-    const client = wordpress_compatible.createClient({
-        url: apiParams.API_URL,
-        username: apiParams.username,
-        password: apiParams.password
-    });
-
-    console.log("starting...");
-    const filter = {
-        // post_type:"", // post,page 文章类型
-        // post_status:"", // publish 文章状态
-        number: 10,// 获取的文章数量，此例中显示的是10篇文章
-        offset: 0,// 从默认顺序里的第几篇文章开始获取，默认是0，就是从头开始，如果要从第二篇，就可以将此参数修改成为1，这个参数适用于文章分列，或者首篇文章不同于其他文章显示
-        orderby: "date", //排序规则，此例为按照时间顺序，默认也是时间顺序
-        // s:"java",
-        order: "desc" // 'ASC'升序，'DESC' 降序
-    };
-    const fields = ["id", "title", "name", "date"];
-    client.getPosts(filter, fields, function (error: any, posts: any) {
-        if (error) {
-            console.log("An error occurred:", error);
-            return;
-        }
-        console.log("Found " + posts.length + " posts!");
-        console.log(posts);
-    });
-}
